+++ src/api/Models/Borehole.cs
@@ -468,6 +468,12 @@ public class Borehole : IChangeTracking, IIdentifyable
     public ICollection<BoreholeFile>? BoreholeFiles { get; set; }
 
     /// <summary>
+    /// Gets the <see cref="Borehole"/>'s photos.
+    /// </summary>
+    [IncludeInExport]
+    public ICollection<Photo>? Photos { get; set; }
+
+    /// <summary>
     /// Gets the <see cref="Codelist"/>s used by the <see cref="Borehole"/> to store multiple id types.
     /// </summary>
     public ICollection<Codelist>? Codelists { get; }